+++ Scripts/Grenade.gd
@@ -46,7 +46,7 @@ func _physics_process(_delta):
 			for i in get_slide_collision_count():
 				var collision = get_slide_collision(i)
 
-				if collision.get_collider().name == "Arena":
+				if collision.get_collider() is TileMapLayer:
 					break
 
 				if collision.get_collider().name == "ExitDoor":