+++ Scripts/SharkSpray.gd
@@ -18,7 +18,7 @@ func _physics_process(_delta):
 	for i in get_slide_collision_count():
 		var collision = get_slide_collision(i)
 
-		if collision.get_collider().name == "Arena":
+		if collision.get_collider() is TileMapLayer:
 			self.queue_free()
 			break